separate outer and central
